Move more stop the clock stuff to the stc package.
